Move Service tests to proper package.
