+++ src/test/java/com/app/car/rental/backend/service/AvisRateServiceTest.java
@@ -1,4 +1,4 @@
-package com.app.car.rental.backend.service.mapper.avis;
+package com.app.car.rental.backend.service;
 
 import com.app.car.rental.backend.service.avis.AvisRateService;
 import com.app.car.rental.backend.web.model.request.RateRequestDto;
