+++ src/test/java/com/app/car/rental/backend/service/AvisReservationRequestServiceTest.java
@@ -1,4 +1,4 @@
-package com.app.car.rental.backend.service.mapper.avis;
+package com.app.car.rental.backend.service;
 
 import com.app.car.rental.backend.api.avis.model.reservation.post.request.*;
 import com.app.car.rental.backend.api.avis.model.reservation.post.response.AvisApiReservationPostResponse;
